app/backend/src/requirements.txt
boto3==1.24.5
flask==2.1.1
werkzeug==2.0.3
PyJWT==2.5.0
flask-cors==6.0.1

